fixed redundant types, log typos, and added null check
